Implemented desync measure all (issue #38) as a general task
