+++ tests/test_integration.py
@@ -8,6 +8,7 @@ from app import create_app, db, app_manager
 
 class IntegrationTestCases(unittest.TestCase):
     def setUp(self):
+        time.sleep(3)
         self.app = create_app('testing')
         self.app_context = self.app.app_context()
         self.app_context.push()
@@ -15,6 +16,7 @@ class IntegrationTestCases(unittest.TestCase):
 
         self.client = self.app.test_client()
         app_manager.dataManager.store_permanent()
+        app_manager.dataManager.last_seen_id = {'values': {}, 'events': {}}
 
     def tearDown(self):
         app_manager.end()
@@ -27,7 +29,7 @@ class IntegrationTestCases(unittest.TestCase):
         ############################################################################
         ############################# start a device ###############################
         ############################################################################
-        
+
         PBR_id = 'PBR-PSI-test01'
         test_PBR_device = {
             'device_id': PBR_id,
@@ -38,11 +40,11 @@ class IntegrationTestCases(unittest.TestCase):
 
         response = self.client.post('http://localhost:5000/device', json=test_PBR_device)
         self.assertTrue(response.json["success"])
-        
+
         ############################################################################
         ########################## start measure task ##############################
         ############################################################################
-        
+
         # settings for periodical measurement task
         measure_all_task_id = 'task-measure-PBR-test-01'
         measure_all_task = {
@@ -57,14 +59,14 @@ class IntegrationTestCases(unittest.TestCase):
                             'upper_tol': 5,
                             'od_attribute': 1
                            }
-        
+
         response = self.client.post('http://localhost:5000/task', json=measure_all_task)
         self.assertTrue(response.json["success"])
-        
+
         ############################################################################
         ############################ start turbidostat #############################
         ############################################################################
-        
+
         # settings for turbidostat task
         pump_task_id = 'task-turbidostat-PBR-test-01'
         turbidostat_task = {
@@ -79,17 +81,17 @@ class IntegrationTestCases(unittest.TestCase):
                             'pump_on_command': {'command_id': '8', 'arguments': '[5, True]'},
                             'pump_off_command': {'command_id': '8', 'arguments': '[5, False]'}
                            }
-        
+
         response = self.client.post('http://localhost:5000/task', json=turbidostat_task)
         self.assertTrue(response.json["success"])
-        
+
         # sleep for some time
         time.sleep(2)
-        
+
         ############################################################################
         ############################## send a command ##############################
         ############################################################################
-        
+
         # send a random command - to change some value
         cmd = {'device_id': PBR_id,
                'command_id': '3',
@@ -97,53 +99,53 @@ class IntegrationTestCases(unittest.TestCase):
                }
         response = self.client.post('http://localhost:5000/command', json=cmd)
         self.assertTrue(response.json["success"])
-        
+
         ############################################################################
         ############################## get all values ##############################
         ############################################################################
-        
+
         # ask for all values
         response = self.client.get('http://localhost:5000/data?device_id={}&type=values'.format(PBR_id))
         self.assertTrue(response.json["success"])
         self.assertTrue(len(response.json["data"]) != 0)
-        
+
         values_keys_first = set(map(int, response.json["data"]))
-        
+
         # sleep for some more time
         time.sleep(20)
-        
+
         ############################################################################
         ############################## get all events ##############################
         ############################################################################
-        
+
         # ask for all events
         all_events = {'device_id': PBR_id,
                       'type': 'events'}
         response = self.client.get('http://localhost:5000/data?device_id={}&type=events'.format(PBR_id))
         self.assertTrue(response.json["success"])
         self.assertTrue(len(response.json["data"]) != 0)
-        
+
         ############################################################################
         ######################## get values from a log ID ##########################
         ############################################################################
-        
+
         # ask again with new log_id to check if got only new ones
         last = max(values_keys_first)
         URL = 'http://localhost:5000/data?device_id={}&type=values&log_id={}'
         response = self.client.get(URL.format(PBR_id, last))
         self.assertTrue(response.json["success"])
         self.assertTrue(len(response.json["data"]) != 0)
-        
+
         values_keys_second = set(map(int, response.json["data"]))
         # check if intersection is empty
         self.assertTrue(values_keys_first & values_keys_second == set())
         # and check all values are newest than the last one
         self.assertTrue(all(list(map(lambda item: item > last, values_keys_second))))
-        
+
         ############################################################################
         ###################### get values from a time point ########################
         ############################################################################
-        
+
         # get data from a time point
         from_time = datetime.utcnow().strftime("%Y%m%d%H%M%S%f")
         time.sleep(15)
@@ -151,62 +153,131 @@ class IntegrationTestCases(unittest.TestCase):
         response = self.client.get(URL.format(PBR_id, from_time))
         self.assertTrue(response.json["success"])
         self.assertTrue(len(response.json["data"]) != 0)
-        
+
         ############################################################################
         ####################### another parallel experiment ########################
         ############################################################################
-        
+
         # settings for test PBR 2
         PBR_id_2 = 'PBR-PSI-test02'
         test_PBR_device['device_id'] = PBR_id_2
-        
+
         response = self.client.post('http://localhost:5000/device', json=test_PBR_device)
         self.assertTrue(response.json["success"])
-        
+
         # measure all for PBR 2
         measure_all_task_id_2 = 'task-measure-PBR-test-02'
         measure_all_task['task_id'] = measure_all_task_id_2
         measure_all_task['device_id'] = PBR_id_2
-        
+
         response = self.client.post('http://localhost:5000/task', json=measure_all_task)
         self.assertTrue(response.json["success"])
-        
+
         # turbidostat for PBR 2
-        
+
         pump_task_id_2 = 'task-turbidostat-PBR-test-02'
         turbidostat_task['task_id'] = pump_task_id_2
         turbidostat_task['device_id'] = PBR_id_2
         turbidostat_task['measure_all_task_id'] = measure_all_task_id_2
-        
+
         response = self.client.post('http://localhost:5000/task', json=turbidostat_task)
         self.assertTrue(response.json["success"])
-        
+
         time.sleep(30)
-        
+
         ############################################################################
         ########################### end tasks and device ###########################
         ############################################################################
-        
+
         # PBR 2
-        
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'task', 'target_id': pump_task_id_2})
         self.assertTrue(response.json["success"])
-        
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'task', 'target_id': measure_all_task_id_2})
         self.assertTrue(response.json["success"])
-        
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'device', 'target_id': PBR_id_2})
         self.assertTrue(response.json["success"])
-        
+
         time.sleep(10)
-        
+
         # PBR 1
-        
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'task', 'target_id': pump_task_id})
         self.assertTrue(response.json["success"])
-        
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'task', 'target_id': measure_all_task_id})
         self.assertTrue(response.json["success"])
-        
+
+        response = self.client.post('http://localhost:5000/end', json={'type': 'device', 'target_id': PBR_id})
+        self.assertTrue(response.json["success"])
+
+    def test_integration_desync(self):
+        ############################################################################
+        ############################# start a device ###############################
+        ############################################################################
+
+        PBR_id = 'PBR-PSI-test01'
+        test_PBR_device = {
+            'device_id': PBR_id,
+            'device_class': 'test',
+            'device_type': 'PBR',
+            'address': 'null',
+        }
+
+        response = self.client.post('http://localhost:5000/device', json=test_PBR_device)
+        self.assertTrue(response.json["success"])
+
+        ############################################################################
+        ###################### start desync measure task ###########################
+        ############################################################################
+
+        frequency_to_commands = {
+            10: {"o2": {"id": "14"}},
+            6: {"ph": {"id": "4", "args": [5, 0]}, "light_1": {"id": "9", "args": [1]}},
+            2: {"od_0": {"id": "5", "args": [0, 30]}}
+        }
+
+        measure_all_task_id = 'task-measure-PBR-test-01'
+        measure_all_desync_task = {
+            "task_id": measure_all_task_id,
+            'task_class': 'General',
+            'task_type': 'measure_all_desync',
+            "device_id": PBR_id,
+            "frequency_to_commands": frequency_to_commands}
+
+        response = self.client.post('http://localhost:5000/task', json=measure_all_desync_task)
+        self.assertTrue(response.json["success"])
+
+        time.sleep(10)
+
+        ############################################################################
+        ######################### check content of DB ##############################
+        ############################################################################
+
+        # ask for all values
+        response = self.client.get('http://localhost:5000/data?device_id={}&type=values'.format(PBR_id))
+        self.assertTrue(response.json["success"])
+        self.assertTrue(len(response.json["data"]) != 0)
+
+        data = response.json["data"]
+
+        o2_count = len(list(filter(lambda item: item['var_id'] == 'o2', data.values())))
+        pH_count = len(list(filter(lambda item: item['var_id'] == 'pH', data.values())))
+        od_count = len(list(filter(lambda item: item['var_id'] == 'od', data.values())))
+
+        self.assertEqual(o2_count, 1)
+        self.assertEqual(pH_count, 2)
+        self.assertTrue(4 < od_count <= 6)
+
+        ############################################################################
+        ########################### end task and device ############################
+        ############################################################################
+
+        response = self.client.post('http://localhost:5000/end',
+                                    json={'type': 'task', 'target_id': measure_all_task_id})
+        self.assertTrue(response.json["success"])
+
         response = self.client.post('http://localhost:5000/end', json={'type': 'device', 'target_id': PBR_id})
         self.assertTrue(response.json["success"])
